MonitorDjango/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bcrypt==4.1.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2024.2.2
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==3.2.25
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-rest-framework==0.1.0
django-simpleui==2023.12.12
djangorestframework==3.15.1
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
gevent==22.10.2
greenlet==3.0.3
importlib-metadata==4.13.0
kombu==5.2.4
numpy==1.21.6
pandas==1.3.5
paramiko==3.4.0
postgres==4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.21
pygrok==1.0.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
regex==2023.12.25
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
zope.event==5.0
zope.interface==6.2
